requirements.txt
streamlit
requests
beautifulsoup4
faiss-cpu
numpy
sentence-transformers
python-dotenv





